feat(web): serve channels UI under /chat/rooms/{roomId}
Remove /channels routes with no redirects per cutover spec.
