+++ apps/web/src/app/(app)/chat/components/channel-draft-shared.tsx
