+++ apps/web/src/app/(app)/chat/components/thread-panel.tsx
